+++ core/src/Environnement/Terrain/Terrain.java
@@ -109,7 +109,7 @@ public class Terrain implements Updatable {
     }
 
     @Override
-    public void update(int delta_t) {
+    public void update(double delta_t) {
         //TODO
     }
 }
